Axios Delete endpoint location changed to Activity Plan Component
